requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
fuzzywuzzy==0.18.0
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
Levenshtein==0.25.0
oauthlib==3.2.2
openai==1.13.3
outcome==1.3.0.post0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQt5-stubs==5.15.6.0
PySocks==1.7.1
python-dotenv==1.0.1
python-Levenshtein==0.25.0
rapidfuzz==3.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
